Add target to program stm32f4-discovery boards
The current openocd now has built-in support built in support for the
stm32f4-discovery board.  So this target does not need the stm32f4x.cfg
config file (and in fact, that file is not correct because the discovery
includes a built in st-link debug device).

To program the PX4 bootloader onto the discovery:
make
make -f Makefile.f4 TARGET=discovery upload-discovery